requirements.txt
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==4.4.2
dill==0.3.8
executing==2.0.1
filelock==3.13.3
fonttools==4.50.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
ipykernel==6.29.4
ipython==8.23.0
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
moviepy==1.0.3
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
# nvidia-cublas-cu12==12.1.3.1
# nvidia-cuda-cupti-cu12==12.1.105
# nvidia-cuda-nvrtc-cu12==12.1.105
# nvidia-cuda-runtime-cu12==12.1.105
# nvidia-cudnn-cu12==8.9.2.26
# nvidia-cufft-cu12==11.0.2.54
# nvidia-curand-cu12==10.3.2.106
# nvidia-cusolver-cu12==11.4.5.107
# nvidia-cusparse-cu12==12.1.0.106
# nvidia-nccl-cu12==2.19.3
# nvidia-nvjitlink-cu12==12.4.99
# nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.1
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
proglog==0.1.10
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
sympy==1.12
thop==0.1.1.post2209072238
torch==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
triton==2.2.0
typing_extensions==4.10.0
tzdata==2024.1
ultralytics==8.1.40
urllib3==2.2.1
wcwidth==0.2.13
